overlord/devicestate: drop undo of gadget update
Drop support for undoing of gadget updates. The system is rebooted during
update, if it boots up to the point where snapd runs we deem the new assets (be
it bootloader or firmware) functional. However, once gadget snap is reverted,
the on-disk state may not be in sync with the description in gadget.yaml.

In other words, all parties taking part in updating of out-of-snap gadget
assets, must be involved in proper quality control of the process and use
available mechanisms (eg. refresh control).

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>