overlord: use private YAML for TestDisconnectUndo
A number of tests reuse a "common" set of YAML documents that grew
organically and have lots of tentacles with quirky properties. An
upcoming change would start to detect those differences. As a step
towards cleaning that up, the various YAMLs are moving towards
individual tests where their content is close to the code using it,
making the test more readable and comprehensible.

This patch adjusts TestDisconnectUndo.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>