seccomp: add 'actlog-supported' command and expose via sandbox-features
snap-seccomp currently queries for ActLog support to decide if it should
be enabled or not. We want to expose this in sandbox-features and also
to other parts of snapd. Because snapd is intended to be compilable with
CGO_ENABLED=0, the simplest solution of refactoring the support check
does not work because it pulls in libseccomp. Therefore, add a new
snap-seccomp command, 'actlog-supported' and have the seccomp backend
call it on initialization so we only call this once on startup instead
of every time the query is performed.