require newer golang-seccomp
Due to https://github.com/seccomp/libseccomp-golang/issues/22,
golang-seccomp <= 0.9.0 cannot create correct BPFs for this feature.
The package does not contain any version information, but we know that
ActLog was implemented in the library after this issue was fixed, so
base the decision on that. ActLog is first available in 0.9.1.