overlord: ensure setup doesn't clobber unrelated state
When a security setup is performed on a single snap then it is
effectively performed on all the apps and hooks of all the snaps that
are connected to the snap that was being set up.

Because internally the setup consists of removing and re-creating
interface connections it will also update stale attributes. This test
ensures that static attributes stored in the connection state of
unrelated snaps are unaffected.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>