cmd/snap-update-ns: use consistent UID during test
In the past the code computing XDG_RUNTIME_DIR would use os.Getuid which
was mirrored by the test. Recently it switched to accepting the UID
argument directly. The test code used constant 1000 which accidentally
matched default user UID on many systems. In Travis, apparently, the UID
running the unit tests is 2000 so the inconsistency caused a failure.
This patch fixes that by using the same UID for the input and the
expected output.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>