snap-seccomp: add 4th field to version-info for golang-seccomp features
After a small refactor to have actLogSupported(), have versionInfo()
call out to it and append the result to the version-info output,
defaulting to '-' with no detected features, then adjust tests.

In sandbox/seccomp/compiler.go, adjust the regex to have field lengths
and drop the fuzzy length check. The lengths could be more specific with
exact lengths for build-id and hash, but retain the previous behavior
and accept a length of one for these. Add a bunch of tests for the
regexp.