snapstate: revert "overlord/snapstate: remove PlugsOnly"
This reverts commit f77fbff18acacd73cdc816812517b90a75d6ee0c.

It turns out that we removed PlugsOnly a bit too early. While
it is not used right now it will be used in the future by a
PR that Samuele is working on. So revert the removal here
to ensure it does not slow the work on this PR down.