daemon, o/snapstate, store: support for installing from cohorts
With this change, the daemon accepts "cohort-key" in the install
instruction when POSTing to /v2/snaps/<snap>, and calls the new
snapstate.InstallCohort with that cohort. This then hits the store
with a CohortKey in the Action, and sets CohortKey on context from
there on.

The bits in snapstate's undoLinkSnap will get covered in refresh,
where we can check the right cohort key is left on undo (the undo test
when it's a fresh install adds nothing new over what exists already).