packaging: use base core to build snap package
Add base core to the snapcraft yaml file to obtain predictable builds
and to use maintained, current versions of snapcraft. (We can also
move snapcraft files out of the way once LP #1827095 is fixed.)

Also with the new snapcraft version .go files in snap/ are not included
in the snap package.

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>