snapcraft.yaml: include libc6 in snapd
We added some tools like xdelta3 and squashfs-tools to the snapd
snap now so that we can use those even if they are not available
on the system. Unfortunately this means that we need to also
carry their libc6 to be truely independent of the host system.

This PR adds libc6 to the snapd snap
We added some tools like xdelta3 and squashfs-tools to the snapd
snap now so that we can use those even if they are not available
on the system. Unfortunately this means that we need to also
carry their libc6 to be truely independent of the host system.

This PR adds libc6 to the snapd snap.