adjust daemon-user test to runtime detect the requirements
At present, no systems meet the requirements for the daemon user
features. We would like to be able to test both when the requirements
are met and when they aren't. Introduce the EXFAIL environment variable
to list systems that we know are expected to fail. This list should only
be reduced as time goes on (eg, fedora-29 is supposed to have everything
soon, as is Ubuntu).

Detect the requirements by parsing the other helper snap's seccomp
profile header. Then:
- requirements not met and EXFAIL - MATCH on install error
- requirements not met and not EXFAIL - error that we regressed
- requirements met and EXFAIL - error that test should be updated
- requirements met and not EXFAIL - install and proceed with tests