daemon, overlord: support for cohort-key in refresh and switch
This implements support for setting the cohort-key of a snap via
refresh or switch.

It also includes a refactor to `snapstate.Update` and `.Switch` that
makes them take an options struct. In fact the majority of the diff is
this refactor.

Right now they are two completely separate structs, but I was
wondering whether instead `snapstate.UpdateOptions` should embed
`snapstate.SwitchOptions`. Maybe it's too "cute", but it feels like it
makes sense, semantically.

Anyway, here goes nothing.