snap-confine: improve error when running on a not /home homedir
We currently don't support running on a non /home homedirs. The
error in this case is very cryptic:
```
cannot create user data directory: /data1/home/username/snap/hello/20: Read-only file system
```

This PR detects the error in snap-confine and provides a clearer
error message to the user with a link to the relevant forum topic.

Ideally we would have a step-by-step guide in the forum docs how
to fix things there.