cmd: rework `snap run --gdb` to work as user
Right now snap run --gdb will run everything as root.
This is a problem because most developers expect to debug
their programs as user. However we need to run as root
because the setuid snap-confine will not run with root
permissions when run under gdb.

The fix is to change snap-gdb-shim to `raise(SIGSTOP)` and
catch that in snap run. Then `snap run` can continue and
snap-gdb-shim calls raise(SIGTRAP) which will drop the user
into the gdb environment.