devicestate: allow remodel to different kernels
This change allows a Remodel() operation to switch to a new
kernel in the new model assertion.

It also adds a spread test that checks that this works for
real. In order to do this a new "test-snapd-pc-kernel" snap
is added to the store that contains a snapshot of an actual
pc-kernel snap for testing.