New hotplug test executed on core16
Now less variables are needed to run nested tests.
When the tests are executed on google-nested:ubuntu-16.04-64, then the
nested vm is automatically set to either core16 or classic xenial, and
when the tests are executed on google-nested:ubuntu-16.04-64, then the
nested vm is automatically set to either core18 or classic bionic.

It is missing the assertion for core 18 as part of the
$TESTSLIB/assertions/ directory.