daemon: only allow `users`/`create-users` when not on classic*
* or an environment variable is set to override it

Without this change, the `/v2/users` and `/v2/create-users` were
always 'live', even on classic. This change makes them bail
immediately if not classic, and an ad-hoc environment variable is not
set.

It also moves all the user-related api bits to `api_users.go` and
their tests to `api_users_test.go`. Hopefully this refactor does not
get in the way.