cmd: add snap-verify stub binary
We will need a snap-verify binary in snapd that is able to validate
snaps from e.g. the recovery partition using a set of assertions
from a given directory.

This PR adds a stub implementation of the tool and makes the
packaging aware of it. For now it will only be available in
the Ubuntu packaging as it is a rather specialized tool. This
may change in the future.