tests: put tests/lib/bin on PATH, move mountinfo-tool
This patch adds a new directory to PATH, allowing test tools to be
written there, without the need to refer to $TESTLIB directly and
without the need to source a shell script to use a function, when that
same logic can be implemented in a standalone program.

At the same time, move the mountinfo-tool program there, for
convenience.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>