Change strace parameters on snap-run test to avoid the test gets stuck
The command 'snap run --strace="-yy"' is getting stuck sporadically.

The command is getting stuck by running (in the strace log):
[pid 11307] 18:30:01.085151 execve("/snap/test-snapd-tools/x1/bin/echo",
["/snap/test-snapd-tools/x1/bin/ec"..., "hello-world"], 0xc000110600 /*
89 vars */ <unfinished ...>

This is not being reproducen on any system, so far it is just happening
on arch linux and opensuse tumbleweed.

The idea of the fix is to change the parameter used to -V and check the
version is printer properly.