devicestate: disallow removal of snaps used in booting early
During the review of #6835 the question about the order of
the checks in canRemove for the active boot snaps was raised.

This PR improves snapsate.canRemove() so that it checks early
if the revision about to be removed is currently used for the
boot. Also adds some unit tests around this.

In practise its a bit of a corner case. A kernel/base used for
booting is usually active and we disallow removing of active
revisions anyway. The window is only when snapd refreshed
base/kernel but has not rebooted yet. But then it does not
accept commands right now until the reboot happend. But even
though we should be careful in canRemove().