tests: replace "! foo | ... | bar" with "not foo | ... | bar"
This patch replaces more complex negated commands, those involving pipes
with the help of the two sed expressions:

    find -name 'task.yaml' -exec sed -i -E -e 's/^( *)! *(.?+) +\| *([^|]+[^\]$)/\1\2 | not \3/' {} \;
    find -name 'task.yaml' -exec sed -i -E -e 's/^( *)! *(.?+) *\| *([^|]+[^\]$)/\1\2 | not \3/' {} \;

The two commands are almost identical, differing in only the approach
towards white-space before the pipe operator. This was done because sed
doesn't natively support non-greedy expressions and I wanted the output
to have consistent spacing.

It is worth noting that the expression excludes commands that have a
continuation on the following line, since those would not be handled
correctly. As a simple example:

    ! foo | bar \
    | froz

The diff was rather uneventful but one replacement result was manually
left out, because it had matched a command like "! foo || bar". Such
exceptions are not worth encoding in the automation since they occur
very infrequently in our code.

Please see the previous commit for a discussion on why pipes need to be
handled this way.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>