httputil: don't retry when network is down
Running our test suite on an offline system, or more conveniently, in a
network namespace with only loopback device, I ran across this error:

    FAIL: retry_test.go:406: retrySuite.TestRetryDoesNotFailForPermanentDNSErrors

    retry_test.go:424:
	// we try exactly once, a non-existing server is a permanent error
	c.Assert(n, Equals, 1)
    ... obtained int = 5
    ... expected int = 1

The retry test would attempt to perform a network operation five times
even though the system has no chance of resolving any names because it
is offline.

Inspection of the returned errors indicates the following error chain:

    outer:  &url.Error{Op:"Get", URL:"http://nonexistingserver909123.com/", Err:(*net.OpError)(0xc000200140)}
    nested: &net.OpError{Op:"dial", Net:"tcp", Source:net.Addr(nil), Addr:net.Addr(nil), Err:(*net.DNSError)(0xc0001b8080)}
    nexted: &net.DNSError{Err:"dial udp 172.16.124.2:53: connect: network is unreachable", Name:"nonexistingserver909123.com", Server:"172.16.124.2:53", IsTimeout:false, IsTemporary:true}

Crucially the innermost error is marked as temporary.

The logic we have first checks if an error is temporary, and if so,
immediately schedules retry, and only then checks if the system is
offline or if DNS is unavailable.

I admit that the retry logic is complex but on a quick look swapping
those checks around is the right thigh to do. When we're offline there's
no point in retrying.

The test now passes and there are no regressions in any systems in
spread.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>