tests: mountinfo-tool --one prints matches on failure
When mountinfo-tool is invoked with --one it expects to see exactly one
mount point matching the filter expression. When more are found an error
is printed but the user is left wondering what really happened. This
patch changes that, so that output is printed unconditionally and the
failure related to using --one is only reported later.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>