cmd/snap-update-ns: don't propagate detaching changes
When snap-update-ns is unmounting something that can host additional
mount points inside, it is now using MNT_DETACH to detach the mount
point (recursively) and lazily unmount anything that is no longer used.
It can be though about like removing a reference in a garbage collected
memory management system. Unlike such systems, however, the kernel also
implements mount and unmount event *propagation*. This can replicate the
change in peer groups that receive propagation events. This can, oddly,
cause unmount loops where the kernel returns EBUSY, because the unmount
operation (even with MNT_DETACH) affects itself, in a way.

One such example is added as a test case in this patch, a fragment of which is
presented here:

    // The mount profile contains but two entries:
    //
    // 1) recursive-bind $SNAP/b/c -> $SNAP/a
    // 2) recursive-bind $SNAP/d -> $SNAP/b
    //
    // NB: The notation "foo -> bar" can be read as "foo is a view of bar",
    // and can be implemented by mount "--rbind bar foo".
    //
    // Both mount operations are performed under a substrate that is MS_SHARED.
    // Therefore, due to the rules that decide upon propagation of bind mounts
    // the propagation of the new mount entries is also shared. This is
    // documented in https://lwn.net/Articles/159092/ (section 5b).
    //
    // Interactive experimentation shows that the following three mount points exist
    // after this operation, as illustrated by findmnt:
    //
    // └─/snap/test-snapd-layout/x1          /dev/loop1     squashfs    ro,nodev,relatime
    //   ├─/snap/test-snapd-layout/x1/b/c    /dev/loop1[/a] squashfs    ro,nodev,relatime
    //   └─/snap/test-snapd-layout/x1/d      /dev/loop1[/b] squashfs    ro,nodev,relatime
    //     └─/snap/test-snapd-layout/x1/d/c  /dev/loop1[/a] squashfs    ro,nodev,relatime
    //
    // Note that after the first mount operation only one mount point is created, namely
    // $SNAP/b/c -> $SNAP/a. The second recursive bind mount not only creates
    // $SNAP/d -> $SNAP/b, but also replicates $SNAP/b/c -> $SNAP/a as
    // $SNAP/d/c -> $SNAP/a.

When this mount profile is undone, the unmount --lazy (aka MNT_DETACH)
operation on x1/d fails with EBUSY, as explained above. The patch solves that
by changing event propagation, in the unmounted tree, to private. This idea is
hinted at in the umount(2) manual page.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>