many: install snap-confine-debug while testing
For the longest time, snap-confine had a sibling, snap-confine-debug,
built with additional debugging messages. Those were almost never visble
outside of a particular contributor's computer. This patch creates a new
binary package, snap-confine-debug, which exists solely to ship the binary
so that it can be injected into the repackaged core or snapd snaps and
installed as a replacement for snap-confine during testing.

The patch is partial since it needs to be compensated in other packaging
systems, so that the new file is not installed anywhere.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>