cmd/snap-confine: umount everything under snap's view of hostfs
We used to unmount the second view of /sys, /dev and /proc that were
visible in /var/lib/snapd/hostfs, because those would routinely upset
certain programs that, for whatever reason, look for sysfs rather than
assuming it is at /sys. This patch extends that idea to unmount everything
visible under /hostfs. This reduces the clutter inside the per-snap
mount namespace.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>