data/selinux: allow running hooks and services from classic snaps
When a classic snap with service or hooks is installed on SELinux system, the
policy will prevent snapd from running the hooks or, should the snap have a
service, it will fail to be started by systemd -> snap-confine chain.

The reason is that for classic snaps, snap-confine executes snap-exec from the
host (labeled as snappy_exec_t) instead of the one from a mounted snap (labeled
as snappy_snap_t). The policy does not allow that.

Allow snappy_exec_t to be an entrypoint to unconfined_service_t.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>