tests: add check for invalid PR titles in the static checks
In the past PRs got merged that did not follow our convention
of having PR titles of the format: "module: summary". Not
following this convention will also make the changelog
generating tool unhappy (it will skip those PRs that do not
follow the conventions).

This PR is a first step to fix this. The "title" check is very
simplistic right now but should be good enough for a start.