many: add strutil.ElliptLeft, use it for shortening cohorts
When I first wrote some cohort code, for 'snap download', I very
carefully pointed out that

    // cohort keys are really long, and the rightmost bit being the
    // interesting bit, so ellipt the rest

and then wrote a private helper

    // TODO: maybe move this to strutil next to ElliptRight

When I then moved to write the rest of the cohort shortening code I
forgot about this and just used ElliptRight (which throws away the
interesting bit, and keeps the bit that changes less).

This makes moves `elliptLeft` to `strutil`, and uses that in all
cohort-shortening code.

HTH, HAND!