tests: rewrite confusing comments
I wrote the initial test before I had the fix so the test used to
observe the old broken behavior. It seems that I left large parts of the
old comments after fixing the test and this made the whole text
self-conflicting. This patch fixes that.

Thanks to Jamie for spotting this.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>