gadget: make raw updater handle shifted structures
A positioned structure has a start-offset that is an absolute value within the
volume. In a system, a structure which has a partition, will correspond to a
single device, thus its actual starting offset for all operations will be 0.

Make raw updater aware of offset returned by device lookup and accordingly
adjust the structure for further I/O operations.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>