snap-confine: move seccomp load after permanent privilege drop (#7032)
* snap-confine: move seccomp load after permanent privilege drop

Because snap-confine would load its seccomp profiles before permanently
dropping privileges, the seccomp profile needed to have rules for
setresuid() (glibc's implementation of seteuid()), setuid() and setgid()
to the dropped to user. Since traditionally we allowed use of all these
syscalls by default (letting AppArmor capability rules and DAC mediate
the access), this ordering has not been an issue. With the introduction
of privilege dropping in snaps, we no longer allow use of the bare
syscalls when the snap specifies it wants to privilege drop, causing
non-root snap commands to fail to run when snap-confine attempted to
drop privileges (this doesn't affect root snap commands since we already
allowed these calls in the privilege dropping seccomp snippet).

One approach to handling this would be for snap-confine to implicitly
add the following rules:

  setuid UID
  setgid GID
  setresuid -1 UID -1

where UID is calculated at runtime as the dropped-to user. Since this
must happen at runtime, snap-seccomp cannot be leveraged and we would
need to either attach a 3rd seccomp filter (the main one, global plus
this new one) or we would need to monkey patch the main profile prior to
loading into the kernel.

A cleaner approach is to instead rearrange the ordering of calls such
that we permanently privilege drop prior to loading the seccomp
profiles. Normally setuid/setgid not only permanently drops the UID/GID,
but also clears the capabilities bounding sets (see "Effect of user ID
changes on capabilities" in 'man capabilities'). To load a seccomp
profile, we need either CAP_SYS_ADMIN or PR_SET_NO_NEW_PRIVS. Since NNP
causes issues with AppArmor and exec transitions in certain snapd
interfaces, we will regain CAP_SYS_ADMIN temporarily after permanently
dropping privileges so we may then load the seccomp profiles, then drop
it after. This provides a small security benefit since we permanently
drop earlier in snap-confine and keep only the capabilities required to
load the seccomp policy.