cmd/libsnap: simplify infofile parser
After some discussions on IRC, in wake of the extra complexity afforded
by the code using the callback and state structures, we've decided to
cut the solution to a smaller one so that the bug fix can land. The
generic solution, which is intended to replace the os-release parser,
will be resurrected separately and undergo separate review.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>