tests: replace various numeric mount options
When comparing mount tables some options, such as number of inodes or
tmpfs size can be a problem as it depends on the amount of memory on the
system. To facilitate testing the --renumber option now handles this
case.

The embedded tests showcase the two instances where I saw this in
practice. It can be easily extended on demand.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>