tests: allow renumbering mount namespace identifiers
The Linux kernel allows to bind-mount /proc/pid/ns/{mnt,...} on top of a
file to preserve a namespace object beyond the death of a process. Snapd
uses this feature extensively.

In the mount table, such preserved namespaces have a specific form:

    2290 2266 0:3 mnt:[4026532909] /run/snapd/ns/xkcd-webserver.mnt rw - nsfs nsfs rw

As one can see the field "mnt:[...]" describes the ... root directory of
the file system that is exposed. Normally this attribute encodes the
subset of the filesystem of the major:minor block device that is visible
at the mount point. This abuse of the attribute is a bit odd, but not
unlike other kernel interfaces.

The number in the brackets is a kernel identifier that is not
deterministic in any way. For the purpose of writing tests that measure
the state of the mount table, we would like to have the option re-write
such numbers into something deterministic.

This patch does exactly that. When --renumber option is used such mount
identifiers are replaced by monotonically increasing numbers.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>