tests: don't preserve size= when rewriting mount tables
When systemd mounts tmpfs it will compute and provide the maximum size
of the filesystem as a fraction of the available memory. This quantity
changes from execution to execution I've seen renumbering "failures" as
the computed quantity for /run/user/0 oscillated between 378180k and
378184k, resulting in either reuse of earlier renumbering of /run or an
allocation of a new value. To avoid this issue replace size= with
constant size=VARIABLE.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>