Manually stop the gvfsd-metadata process
seems to be a race between the gvfsd-metadata which is being stopped and
the rm command. By using pkill we make sure the process is finished and
the dir can be removed.

This is the log:

++ ps -ef
...
root     21163     1  0 13:52 ?        00:00:00
/usr/lib/gvfs/gvfsd-metadata
...

++ pidof gvfsd-metadata
+ pid=
+ rm -rf
/home/gopath/src/github.com/snapcore/snapd/tests/main/interfaces-calendar-service/xdg
rm: cannot remove
'/home/gopath/src/github.com/snapcore/snapd/tests/main/interfaces-calendar-service/xdg/share/gvfs-metadata':
Directory not empty