overlord/patch: simplify conditions for re-applying sublevel patches for level 6 (#7016)
Simplify conditions for re-applying sublevel patches for patch level 6: check snapd version string and reset sublevel back to 0 if it's different from last snapd version. This fixes issue with coming from and old core to a new snapd (either from a deb or snapd snap), and with a missing current symlink on core observed in some edge cases of failed rollback.