tests: remove lxd / lxcfs if pre-installed
The new mount namespace test has data sets for google cloud images and
qemu images. Apart from some cgroup differences the main difference is
presence and absence of pre-installed LXD, including the lxcfs
filesystem.

Because of how our tests prepare and restore (in a non-obvious way) and
because chasing this now is ineffective, to ensure the mount namespace
looks the same across test execution, remove lxd if it is installed in
the initial image.

Tests that require lxd install and remove it already.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>