recovery: use luks master key
Change LUKS device operations to use the master key directly. Cryptsetup
requires the master key to be read from the filesystem so we create a
temporary file with the unsealed master key on a RAM-backed device,
however we should consider writing a custom utility to read the key from
stdin instead.

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>