osutil/user.go: add UserGroupAdd()
While the AddUser() function already exists, 'adduser' is a
Debian(-derivative)-specific command and not portable cross-distro, so
introduce the new, UserGroupAdd() which uses useradd and groupadd from
the industry-standard shadow package. Since this function should create
a user and group on the host, we want to use the host's tools to ensure
the user and group are created in a manner consistent with the host
system (eg, in case the distro patches useradd/groupadd for the distro,
we don't want to reexec our own tools).

UserGroupAdd(name, id) will call:

$ groupadd --system --gid <id> [--extrausers] <name>
$ useradd --system --home-dir /nonexistent --no-create-home \
          --shell /bin/false --gid <id> --no-user-group --uid <id> \
          [--extrausers] <name>

to create a no-login, locked system user and group with the specified
{user,group}name and {u,g}id. Note that we must use groupadd instead of
useradd --user-group because --user-group will dynamically pick a gid
from the range defined in login.defs instead of using what is passed to
--uid.