snapstate: make progress reporting less granular
The progress reporting for e.g. downloads is a bit excessive. Each
taskProgressAdapter.Write() will lock the state and update the
task progress. While this is great for the UI it puts strain onto
slow systems and make the download slower. This PR will change the
code to only update the progress if it changed more than 0.2%.