tests: add mountinfo-tool --ref-x1000
When looking at mount tables from the host and the per-snap or per-user
mount namespace it is nice to be easily able to distinguish numeric
identifiers as belonging to one or the other. In addition, if a new
mount entry is added or removed in one namespace it has cascading
effects on renumbered entries in derivative mount namespaces.

To combat both of those problems add the --ref-x1000 option. When used,
each time --ref is used to get deterministic renumbering and comparison
across mount tables belonging to different mount namespaces, the numeric
value of the mount_id, parent_id and peer group number is using ranges
0-999 for the host mount namespace, 1000-1999 for the per-snap mount
namespace and 2000-2999 for the per-user mount namespace.

This has nice consequence of really easy visual analysis of the mount
table propagation settings, since the peer group clearly belongs to a
given set.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>